+++ cppad/local/div_op.hpp
@@ -97,8 +97,8 @@ inline void forward_divvv_op_dir(
 
 	// Taylor coefficients corresponding to arguments and result
 	size_t num_taylor_per_var = (cap_order-1) * r + 1;
-	Base* x = taylor + arg[0] * num_taylor_per_var;
-	Base* y = taylor + arg[1] * num_taylor_per_var;
+	Base* x = taylor + size_t(arg[0]) * num_taylor_per_var;
+	Base* y = taylor + size_t(arg[1]) * num_taylor_per_var;
 	Base* z = taylor + i_z    * num_taylor_per_var;
 
 
@@ -291,7 +291,7 @@ inline void forward_divpv_op_dir(
 
 	// Taylor coefficients corresponding to arguments and result
 	size_t num_taylor_per_var = (cap_order-1) * r + 1;
-	Base* y = taylor + arg[1] * num_taylor_per_var;
+	Base* y = taylor + size_t(arg[1]) * num_taylor_per_var;
 	Base* z = taylor + i_z    * num_taylor_per_var;
 
 	// Using CondExp, it can make sense to divide by zero,
@@ -472,7 +472,7 @@ inline void forward_divvp_op_dir(
 
 	// Taylor coefficients corresponding to arguments and result
 	size_t num_taylor_per_var = (cap_order-1) * r + 1;
-	Base* x = taylor + arg[0] * num_taylor_per_var;
+	Base* x = taylor + size_t(arg[0]) * num_taylor_per_var;
 	Base* z = taylor +    i_z * num_taylor_per_var;
 
 	// Parameter value
